added black file; finished breaking up original checks across modules and placeholder tests
