+++ nwbinspector/__init__.py
@@ -1,1 +1,1 @@
-"""Authors: Cody Baker and Ben Dichter."""from .utils import available_checksfrom .checks.check_nwb_containers import *from .checks.check_time_series import *from .checks.check_nwbfile_metadata import *
+from .utils import available_checksfrom .checks.check_nwb_containers import *from .checks.check_time_series import *from .checks.check_nwbfile_metadata import *
