+++ nwbinspector/checks/check_tables.py
@@ -1,43 +1,87 @@
-"""Authors: Cody Baker and Ben Dichter."""
+"""Authors: Cody Baker, Ben Dichter, and Ryan Ly."""
 import numpy as np
 
 import pynwb
+import hdmf
 
-from ..utils import nwbinspector_check, check_regular_series
-
-
-@nwbinspector_check(severity=2, neurodata_type=pynwb.TimeSeries)
-def check_regular_timestamps(time_series: pynwb.TimeSeries, time_tol_decimals=9):
-    """If the TimeSeries uses timestamps, check if they are regular (i.e., they have a constant rate)."""
-    if time_series.timestamps is not None and check_regular_series(
-        series=time_series.timestamps, tolerance_decimals=time_tol_decimals
-    ):
-        return (
-            "TimeSeries appears to have a constant sampling rate. "
-            f"Consider specifying starting_time={time_series.timestamps[0]} "
-            f"and rate={time_series.timestamps[1] - time_series.timestamps[0]} instead of timestamps."
-        )
+from ..tools import all_of_type
+from ..utils import nwbinspector_check
 
 
-@nwbinspector_check(severity=2, neurodata_type=pynwb.TimeSeries)
-def check_data_orientation(time_series: pynwb.TimeSeries):
-    """If the TimeSeries has data, check if the longest axis (almost always time) is also the zero-axis."""
-    if time_series.data is not None and any(
-        np.array(time_series.data.shape[1:]) > time_series.data.shape[0]
-    ):
-        return (
-            "Data orientation may be in the wrong orientation. "
-            "Time should be in the first dimension, and is usually the longest dimension. "
-            "Here, another dimension is longer. "
-        )
+@nwbinspector_check(severity=3, neurodata_type=pynwb.core.DynamicTable)
+def check_empty_tables(nwbfile):
+    """Check if DynamicTable is empty."""
+    for tab in all_of_type(nwbfile, pynwb.core.DynamicTable):
+        if len(tab.id) == 0:
+            print("NOTE: '%s' %s has no rows" % (tab.name, type(tab).__name__))
+            continue
+        if len(tab.id) == 1:
+            print("NOTE: '%s' %s has one row" % (tab.name, type(tab).__name__))
+            continue
 
 
-@nwbinspector_check(severity=3, neurodata_type=pynwb.TimeSeries)
-def check_timestamps_match_first_dimension(time_series: pynwb.TimeSeries):
-    """If the TimeSeries has timestamps, check if their length is the same as the zero-axis of data."""
-    if (
-        time_series.data is not None
-        and time_series.timestamps is not None
-        and time_series.data.shape[0] != len(time_series.timestamps)
-    ):
-        return "The length of the first dimension of data does not match the length of timestamps."
+@nwbinspector_check(severity=1, neurodata_type=pynwb.core.DynamicTable)
+def check_single_tables(nwbfile):
+    """Check if DynamicTable has only a single row; may be better represented by another data type."""
+    for tab in all_of_type(nwbfile, pynwb.core.DynamicTable):
+        if len(tab.id) == 1:
+            print("NOTE: '%s' %s has one row" % (tab.name, type(tab).__name__))
+            continue
+
+
+@nwbinspector_check(severity=1, neurodata_type=pynwb.core.DynamicTable)
+def check_column_data_is_not_none(nwbfile):
+    """Check column values in DynamicTable to enssure they are not None."""
+    for tab in all_of_type(nwbfile, pynwb.core.DynamicTable):
+        for col in tab.columns:
+            if not isinstance(col, hdmf.common.table.DynamicTableRegion) and col.data is None:
+                return f"'{tab.name}' {type(tab).__name__} column {col.name} data is None"
+                # continue
+                # TODO: think about how to handle this continuable logic in new refactor
+
+
+# TODO, continue to break this up
+@nwbinspector_check(severity=1, neurodata_type=pynwb.core.DynamicTable)
+def check_column_table(nwbfile):
+    """Check column values in DynamicTable objects"""
+    for tab in all_of_type(nwbfile, pynwb.core.DynamicTable):
+        for col in tab.columns:
+            if isinstance(col, hdmf.common.table.DynamicTableRegion):
+                continue
+
+            if col.name.endswith("index"):  # skip index columns
+                continue
+
+            if isinstance(col.data, hdmf.backends.hdf5.h5_utils.DatasetOfReferences):  # TODO find a better way?
+                continue
+
+            uniq = np.unique(col.data)
+            # TODO only do this for optional columns
+            if len(uniq) == 1:
+                error_code = "A101"
+                print(
+                    "- %s: '%s' %s column '%s' data has all values = %s"
+                    % (error_code, tab.name, type(tab).__name__, col.name, uniq[0])
+                )
+            elif np.array_equal(uniq, [0.0, 1.0]):
+                if col.data.dtype.type != np.bool_:
+                    error_code = "A101"
+                    print(
+                        "- %s: '%s' %s column '%s' data should be type boolean instead of %s"
+                        % (
+                            error_code,
+                            tab.name,
+                            type(tab).__name__,
+                            col.name,
+                            col.data.dtype,
+                        )
+                    )
+            elif len(uniq) == 2:
+                error_code = "A101"
+                print(
+                    (
+                        "- %s: '%s' %s column '%s' data has only unique values %s. Consider storing the data "
+                        "as boolean."
+                    )
+                    % (error_code, tab.name, type(tab).__name__, col.name, uniq)
+                )
