+++ tests/unit_tests/test_ophys.py
@@ -1,0 +1,7 @@
+"""Authors: Cody Baker and Ben Dichter."""
+import pytest
+
+
+@pytest.mark.skip(msg="TODO")
+def test_check_ophys():
+    pass
