+++ tests/unit_tests/test_tables.py
@@ -1,0 +1,17 @@
+"""Authors: Cody Baker and Ben Dichter."""
+import pytest
+
+
+@pytest.mark.skip(msg="TODO")
+def test_check_empty_tables():
+    pass
+
+
+@pytest.mark.skip(msg="TODO")
+def test_check_single_tables():
+    pass
+
+
+@pytest.mark.skip(msg="TODO")
+def test_check_column_data_is_not_none():
+    pass
